Delegate to ResponseBodyWriter where possible (#562)
Since #539, all bodies are wrapped in a GenericEntity and must go through JaxRsGenericEntityMessageBodyWriter, which then delegates to the appropriate writer implementation for the type. However some of these writers do not support the full MessageBodyWriter interface and prefer to be called through the newer ResponseBodyWriter API, e.g. because they want to send a `not modified` response when appropriate.

This patch implements ResponseBodyWriter in JaxRsGenericEntityMessageBodyWriter. Since JAX-RS write interceptors potentially need to wrap the OutputStream, the interception code is adapted to transform the ByteBody to streams on demand by buffering the content. If there are no interceptors or if they don't access the OutputStream, the ByteBody or even ByteBodyHttpResponse is returned unmodified, which allows for streaming.

Fixes #561